check for device relevance for all subroutes after device selection
