+++ src/renderer/src/routes/app/projects/$projectId_/invite/devices/$deviceId/route.tsx
@@ -1,0 +1,33 @@
+import { Outlet, createFileRoute, redirect } from '@tanstack/react-router'
+
+import { GenericRoutePendingComponent } from '../../../../../../../components/generic-route-pending-component'
+
+export const Route = createFileRoute(
+	'/app/projects/$projectId_/invite/devices/$deviceId',
+)({
+	// NOTE: We want to make sure that the device of interest is actually still detected by us and connected.
+	// If it's not, then we redirect to the page that lists the discovered devices.
+	beforeLoad: async ({ context, params }) => {
+		const { clientApi } = context
+		const { deviceId, projectId } = params
+		const peers = await clientApi.listLocalPeers()
+
+		const matchingPeer = peers.find((p) => p.deviceId === deviceId)
+
+		if (!matchingPeer || matchingPeer.status === 'disconnected') {
+			throw redirect({
+				to: '/app/projects/$projectId/invite/devices',
+				params: { projectId },
+				replace: true,
+			})
+		}
+
+		return { peerOnLoad: matchingPeer }
+	},
+	pendingComponent: GenericRoutePendingComponent,
+	component: RouteComponent,
+})
+
+function RouteComponent() {
+	return <Outlet />
+}
